Set up bot-blocking for myinvestkit.com root and add error handling to partner home page
Bots are no longer allowed to crawl the root directory (but all others are available) on myinvestkit. This
code is not currently active (the disallow command is still in effect).

Partner home page now has error handling to only display information that is returned rather than requiring
all information before any is displayed.